Home Page status(20%)
